made watchdog look for keepalive file

+++ nmeachecker.sh
@@ -3,6 +3,10 @@
 # Written with the rather erratic help of Gemini
 # Could this be triggering erroneously, in that the file is updated, but the timestamp is not showing correctly ?
 
+
+# This checks for recent file updates in nmea_data/2020-*/*
+# it does not check for the file touch in nmea_logs
+
 # Define threshold (in minutes)
 threshold=13
 usual=1
@@ -44,15 +48,28 @@ for directory in $dir_root; do
     done
 done
 
+stillalive=0 # flag to see if the heartbeat still_alive.txt is still alive =0 means alive
+alivepath="$/root/nmea_logs/still_alive.txt"
+if [ -f "$alivepath" ]; then
+    alive_mtime=$(date -r "$alivepath" +%s)
+    alive_stamp=$(date -r "$alivepath" +"%T %Z")
+
+     if [ $alive_mtime -gt $threshold_time ]; then
+         stillalive=1 # means it is dead
+     fi
+fi
 # Handle results
 if [ $overdue_updated -ne 1 ]; then
-  echo `date` "No files in '$directory' have been updated in the last $threshold minutes."
-  # so kill the .py process, which terminates the .sh script
-  # cron will then restart it in 3 minutes
-  touch /root/nmea_logs/nmealogger-hung.txt
-  echo `date` "Hung: no update in $threshold minutes.  $updated $file_stamp nmeachecker.sh" >> /root/nmea_logs/nmealogger_error.txt
-  pkill -ef "python /root/nmea_gps/nmealogger.py"
-  exit 1
+  if [$stillalive -ne 0 ] ; then
+      # so kill the .py process, which terminates the .sh script
+      # cron will then restart it in 3 minutes
+      touch /root/nmea_logs/nmealogger-hung.txt
+      echo `date` "Hung: no update in $threshold minutes.  $updated $file_stamp nmeachecker.sh" >> /root/nmea_logs/nmealogger_error.txt
+      pkill -ef "python /root/nmea_gps/nmealogger.py"
+      exit 1
+  else
+      echo `date` "Still alive: $alive_stamp even though no recent updates in $directory."
+  fi
 # else
   # echo `date` "$filename updated recently"
 fi
